% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_lm.R
\name{check_autocorrelation}
\alias{check_autocorrelation}
\title{Check Autocorrelation}
\usage{
check_autocorrelation(model)
}
\arguments{
\item{model}{A fitted lm object.}
}
\value{
An htest object or NA if computation fails.
}
\description{
Performs the Durbin-Watson test for first-order autocorrelation in residuals.
}
\examples{
model <- lm(mpg ~ wt + hp, data = mtcars)
check_autocorrelation(model)
}
