% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_lm.R
\name{check_box_tidwell}
\alias{check_box_tidwell}
\title{Check Logistic Linearity with Box-Tidwell Test}
\usage{
check_box_tidwell(model)
}
\arguments{
\item{model}{A fitted glm object.}
}
\value{
A Box-Tidwell test result or NA if computation fails.
}
\description{
Performs a Box-Tidwell test for linearity of continuous predictors in the
logit.
}
\examples{
model <- glm(am ~ wt + hp, data = mtcars, family = binomial)
check_box_tidwell(model)
}
