% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_lm.R
\name{check_heteroskedasticity}
\alias{check_heteroskedasticity}
\title{Check Heteroskedasticity}
\usage{
check_heteroskedasticity(model)
}
\arguments{
\item{model}{A fitted lm object.}
}
\value{
An htest object or NA if computation fails.
}
\description{
Performs Breusch-Pagan test for heteroskedasticity.
}
