% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_lm.R
\name{check_hosmer_lemeshow}
\alias{check_hosmer_lemeshow}
\title{Check Logistic Goodness of Fit}
\usage{
check_hosmer_lemeshow(model)
}
\arguments{
\item{model}{A fitted binomial glm object.}
}
\value{
An htest object or NA if computation fails or the model is not
  binomial.
}
\description{
Performs the Hosmer-Lemeshow goodness-of-fit test for binomial glm models.
}
\examples{
model <- glm(am ~ wt + hp, data = mtcars, family = binomial)
check_hosmer_lemeshow(model)
}
