% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_lm.R
\name{check_normality}
\alias{check_normality}
\title{Check Normality of Residuals}
\usage{
check_normality(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
An htest object or NA if computation fails.
}
\description{
Performs the Shapiro-Wilk test on model residuals.
}
\examples{
model <- lm(mpg ~ wt + hp, data = mtcars)
check_normality(model)
}
