% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_lm.R
\name{check_vif}
\alias{check_vif}
\title{Check Variance Inflation Factors}
\usage{
check_vif(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
A list describing whether VIF computation succeeded. Successful
  results include a `vif` element containing the variance inflation factors.
}
\description{
Computes variance inflation factors to detect multicollinearity.
}
\examples{
model <- lm(mpg ~ wt + hp + disp, data = mtcars)
check_vif(model)
}
