% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_lm.R, R/diagnose_model.R
\name{diagnose_model.glm}
\alias{diagnose_model.glm}
\alias{diagnose_model.lm}
\alias{diagnose_model.coxph}
\alias{diagnose_model}
\title{Diagnose Statistical Models}
\usage{
\method{diagnose_model}{glm}(model, ...)

\method{diagnose_model}{lm}(model, ...)

\method{diagnose_model}{coxph}(model, ...)

diagnose_model(model, ...)
}
\arguments{
\item{model}{A fitted model object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
An object of class "model_diagnostics" containing the results of various diagnostic tests.
}
\description{
This is a generic function for performing diagnostic checks on statistical models.
It dispatches to specific methods based on the model type.
}
\examples{
# Linear model diagnostics
model_lm <- lm(mpg ~ wt + hp, data = mtcars)
diag_lm <- diagnose_model(model_lm)
summary(diag_lm)
plot(diag_lm)

# Logistic regression diagnostics
model_glm <- glm(am ~ wt + hp, data = mtcars, family = binomial)
diag_glm <- diagnose_model(model_glm)
summary(diag_glm)

# Poisson regression diagnostics
model_pois <- glm(carb ~ wt + hp, data = mtcars, family = poisson)
diag_pois <- diagnose_model(model_pois)
summary(diag_pois)

# Cox proportional hazards diagnostics
library(survival)
data(lung)
model_cox <- coxph(Surv(time, status) ~ age + sex + ph.ecog, data = lung)
diag_cox <- diagnose_model(model_cox)
summary(diag_cox)
}
