% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.model_diagnostics}
\alias{print.model_diagnostics}
\title{Print Model Diagnostics}
\usage{
\method{print}{model_diagnostics}(x, ...)
}
\arguments{
\item{x}{An object of class "model_diagnostics".}

\item{...}{Additional arguments passed to print.}
}
\value{
The object x, invisibly.
}
\description{
Prints a summary of the model diagnostics object.
}
